Update for SEQUEL ORM
Add base.rb
Add sequel.rb
Add yaml.rb

Remove helper methods in pixel.rb
Comment out stuff in device.erb
