+++ Gemfile
@@ -4,4 +4,6 @@ ruby '2.0.0'
 gem 'sinatra'
 gem 'sinatra-contrib'
 gem 'pg'
+gem 'sequel'
+gem 'mysql'
