+++ views/device.erb
@@ -1,6 +1,6 @@
-<h1><%= #@device %></h1>
+<h1><%## #@device %></h1>
 <div>
-  <% if !interfaces.empty? %>
+  <%# if !interfaces.empty? %>
   <table id='ifTable' class='table table-hover table-condensed tablesorter'>
     <thead>
       <tr>
@@ -14,28 +14,28 @@
       </tr>
     </thead>
     <tbody>
-    <% interfaces.each do |if_index,oids| %>
-      <tr <%= tr_attributes(oids, { hide_if_child: true, hl_relation: true }) %>>
-        <td class='pxl-hidden'><%= total_bps_cell(interfaces, oids) %></td>
-        <td class='pxl-td-shrink'><%= interface_link(oids) %></td>
-        <td><table class='pxl-linkUp pxl-linkUp-<%= link_status_color(interfaces, oids) %>'><tr>
-              <td data-toggle='tooltip' data-container='body' title='<%= link_status_tooltip(interfaces, oids) %>' data-rel='tooltip-right'></td>
+    <%# interfaces.each do |if_index,oids| %>
+      <tr <%## tr_attributes(oids, { hide_if_child: true, hl_relation: true }) %>>
+        <td class='pxl-hidden'><%## total_bps_cell(interfaces, oids) %></td>
+        <td class='pxl-td-shrink'><%## interface_link(oids) %></td>
+        <td><table class='pxl-linkUp pxl-linkUp-<%## link_status_color(interfaces, oids) %>'><tr>
+              <td data-toggle='tooltip' data-container='body' title='<%## link_status_tooltip(interfaces, oids) %>' data-rel='tooltip-right'></td>
           </tr></table></td>
-        <td class='<%= 'pxl-child-td' if oids['isChild'] %>'>
-          <%= oids['linkType'] %>
-          <% if oids['isParent'] %>
-            <button type='button' class='swapPlusMinus btn btn-default btn-xs' data-toggle='collapse' data-target='.<%= if_index%>_child'>
+        <td class='<%## 'pxl-child-td' if oids['isChild'] %>'>
+          <%## oids['linkType'] %>
+          <%# if oids['isParent'] %>
+            <button type='button' class='swapPlusMinus btn btn-default btn-xs' data-toggle='collapse' data-target='.<%## if_index%>_child'>
               <span class="glyphicon glyphicon-plus"></span>
             </button>
-          <% elsif oids['isChild'] %>
-            <span class='pxl-hidden'><%= "[#{oids['myParentName']}]" %></span>
-          <% end %>
+          <%# elsif oids['isChild'] %>
+            <span class='pxl-hidden'><%## "[#{oids['myParentName']}]" %></span>
+          <%# end %>
         </td>
-        <td><%= neighbor_link(oids) %></td>
-        <td class='pxl-histogram' data-pxl-meta='<%= oids['bpsIn_util'] %>'><%= bps_cell('In', oids) %></td>
-        <td class='pxl-histogram' data-pxl-meta='<%= oids['bpsOut_util'] %>'><%= bps_cell('Out', oids) %></td>
-    <% end %>
+        <td><%## neighbor_link(oids) %></td>
+        <td class='pxl-histogram' data-pxl-meta='<%## oids['bpsIn_util'] %>'><%## bps_cell('In', oids) %></td>
+        <td class='pxl-histogram' data-pxl-meta='<%## oids['bpsOut_util'] %>'><%## bps_cell('Out', oids) %></td>
+    <%# end %>
     </tbody>
   </table>
-  <% end %>
+  <%# end %>
 </div>
